requirements.txt
accelerate==1.10.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
anyio==4.11.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
async-timeout==5.0.1
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.14.2
bitsandbytes==0.39.0
bleach==6.2.0
blinker==1.9.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
comm==0.2.3
contourpy==1.3.2
cycler==0.12.1
datasets==2.12.0
dbus-python==1.2.18
debugpy==1.8.17
decorator==5.2.1
defusedxml==0.7.1
dill==0.3.6
distro==1.7.0
einops==0.6.1
exceptiongroup==1.3.0
executing==2.2.1
fastjsonschema==2.21.2
filelock==3.13.1
Flask==3.1.2
fonttools==4.60.1
fqdn==1.5.1
frozenlist==1.8.0
fsspec==2024.6.1
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.35.3
idna==3.10
ipykernel==6.30.1
ipython==8.37.0
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
joblib==1.5.2
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kiwisolver==1.4.9
lark==1.3.0
loralib==0.1.1
MarkupSafe==2.1.5
matplotlib==3.10.7
matplotlib-inline==0.1.7
mistune==3.1.4
mpmath==1.3.0
multidict==6.6.4
multiprocess==0.70.14
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
notebook_shim==0.2.4
numpy==2.1.2
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.21.5
nvidia-nvtx-cu11==11.8.86
opencv-python==4.12.0.88
outcome==1.3.0.post0
overrides==7.7.0
packaging==25.0
pandas==2.3.3
pandocfilters==1.5.1
parso==0.8.5
peft==0.17.1
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.4.0
prometheus_client==0.23.1
prompt_toolkit==3.0.52
propcache==0.4.0
psutil==7.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==21.0.0
pycparser==2.23
pydicom==3.0.1
Pygments==2.19.2
PyGObject==3.42.1
pyparsing==3.2.5
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-json-logger==4.0.0
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.36.2
regex==2025.9.18
requests==2.32.5
responses==0.18.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rpds-py==0.27.1
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.15.3
seaborn==0.13.2
selenium==4.36.0
Send2Trash==1.8.3
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
ssh-import-id==5.11
stack-data==0.6.3
sympy==1.13.3
terminado==0.18.1
threadpoolctl==3.6.0
tinycss2==1.4.0
tokenizers==0.13.3
tomli==2.2.1
torch==2.7.1+cu118
torchaudio==2.7.1+cu118
torchvision==0.22.1+cu118
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.30.1
trio==0.31.0
trio-websocket==0.12.2
triton==3.3.1
types-python-dateutil==2.9.0.20250822
typing_extensions==4.15.0
tzdata==2025.2
uri-template==1.3.0
urllib3==2.5.0
wcwidth==0.2.14
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
wget==3.2
wsproto==1.2.0
xxhash==3.6.0
yarl==1.21.0
